# BFGS: Add simple line search with Wolfe conditions