Adds a gradient norm-based termination condition for BFGS and utilizes a backtracking line-search that satisfies the Wolfe conditions. Constants C1 and C2 are taken from the suggestion in Nocedal and Wright.

Note that there are better/smarter line search algorithms (see Algorithm 3.5 in Nocedal and Wright, in particular). But function/gradient evaluations are fairly cheap here. Further benchmarking will be needed to determine if it's worth using more advanced step selection.